#36 fix: MainPage 하단 바별 파일 나누기
